# perf: Improve performance of safeParse in case of malicious payload