--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Install Dependencies
         run: npm install
       - name: Test
-        run: npm run test-in-browsers
+        run: npm run test:browsers
   test:
     runs-on: ubuntu-latest
     strategy:
@@ -56,7 +56,7 @@ jobs:
       - name: Install Dependencies
         run: npm install --ignore-scripts
       - name: Test
-        run: npm run test-legacy
+        run: npm run test:legacy
   automerge:
     needs:
       - browsers