--- a/benchmarks/package.json
+++ b/benchmarks/package.json
@@ -2,11 +2,12 @@
   "name": "benchmarks",
   "version": "1.0.0",
   "scripts": {
+    "valid": "node valid.js",
     "ignore": "node ignore.js",
     "no_proto": "node no__proto__.js",
     "remove": "node remove.js",
     "throw": "node throw.js",
-    "all": "node --version && npm run ignore && npm run no_proto && npm run remove && npm run throw"
+    "all": "node --version && npm run ignore && npm run no_proto && npm run remove && npm run throw && npm run valid"
   },
   "dependencies": {
     "benchmark": "^2.1.4"