--- a/benchmarks/remove.js
+++ b/benchmarks/remove.js
@@ -13,8 +13,11 @@ suite
   .add('JSON.parse', () => {
     JSON.parse(internals.text)
   })
-  .add('secure-json-parse', () => {
+  .add('secure-json-parse parse', () => {
     sjson.parse(internals.text, { protoAction: 'remove' })
+  })
+  .add('secure-json-parse safeParse', () => {
+    sjson.safeParse(internals.text)
   })
   .add('reviver', () => {
     JSON.parse(internals.text, internals.reviver)