--- a/benchmarks/valid.js
+++ b/benchmarks/valid.js
@@ -4,8 +4,7 @@ const Benchmark = require('benchmark')
 const sjson = require('..')
 
 const internals = {
-  text: '{ "a": 5, "b": 6, "__proto__": { "x": 7 }, "c": { "d": 0, "e": "text", "__proto__": { "y": 8 }, "f": { "g": 2 } } }',
-  invalid: '{ "a": 5, "b": 6, "__proto__": { "x": 7 }, "c": { "d": 0, "e": "text", "__proto__": { "y": 8 }, "f": { "g": 2 } } } }'
+  text: '{ "a": 5, "b": 6, "c": { "d": 0, "e": "text", "f": { "g": 2 } } }'
 }
 
 const suite = new Benchmark.Suite()
@@ -14,23 +13,14 @@ suite
   .add('JSON.parse', () => {
     JSON.parse(internals.text)
   })
-  .add('JSON.parse error', () => {
-    try {
-      JSON.parse(internals.invalid)
-    } catch (ignoreErr) { }
-  })
   .add('secure-json-parse parse', () => {
-    try {
-      sjson.parse(internals.text)
-    } catch (ignoreErr) { }
+    sjson.parse(internals.text)
   })
   .add('secure-json-parse safeParse', () => {
     sjson.safeParse(internals.text)
   })
-  .add('reviver', () => {
-    try {
-      JSON.parse(internals.text, internals.reviver)
-    } catch (ignoreErr) { }
+  .add('JSON.parse reviver', () => {
+    JSON.parse(internals.text, internals.reviver)
   })
   .on('cycle', (event) => {
     console.log(String(event.target))
@@ -42,7 +32,7 @@ suite
 
 internals.reviver = function (key, value) {
   if (key === '__proto__') {
-    throw new Error('kaboom')
+    return undefined
   }
 
   return value