--- a/benchmarks/valid.js
+++ b/benchmarks/valid.js
@@ -5,7 +5,7 @@ const sjson = require('..')
 
 const internals = {
   text: '{ "a": 5, "b": 6, "c": { "d": 0, "e": "text", "f": { "g": 2 } } }',
-  proto: '{ "a": 5, "b": 6, "__proto__": { "x": 7 }, "c": { "d": 0, "e": "text", "__proto__": { "y": 8 }, "f": { "g": 2 } } }',
+  proto: '{ "a": 5, "b": 6, "__proto__": { "x": 7 }, "c": { "d": 0, "e": "text", "__proto__": { "y": 8 }, "f": { "g": 2 } } }'
 }
 
 const suite = new Benchmark.Suite()
@@ -21,7 +21,7 @@ suite
     sjson.parse(internals.text)
   })
   .add('secure-json-parse parse proto', () => {
-    sjson.parse(internals.text, { constructorAction: 'ignore', protoAction: 'ignore'})
+    sjson.parse(internals.text, { constructorAction: 'ignore', protoAction: 'ignore' })
   })
   .add('secure-json-parse safeParse', () => {
     sjson.safeParse(internals.text)