--- a/index.test-d.ts
+++ b/index.test-d.ts
@@ -1,8 +1,10 @@
 import { expectType, expectError } from 'tsd'
-import sjson = require('.')
+import sjson from "."
 
 expectError(sjson.parse(null))
 expectType<any>(sjson.parse('{"anything":0}'))
+sjson.parse('"test"', { constructorAction: 'error' })
+sjson.parse('"test"', null, { constructorAction: 'error' })
 
 sjson.parse('"test"', null, { protoAction: 'remove' })
 expectError(sjson.parse('"test"', null, { protoAction: 'incorrect' }))
@@ -12,11 +14,6 @@ expectError(sjson.parse('"test"', null, { constructorAction: 'incorrect' }))
 sjson.safeParse('"test"', null)
 sjson.safeParse('"test"')
 expectError(sjson.safeParse(null))
-
-sjson.scan('"test"', { protoAction: 'remove' })
-expectError(sjson.scan('"test"', { protoAction: 'ignore' }))
-sjson.scan('"test"', { constructorAction: 'error' })
-expectError(sjson.scan('"test"', { constructorAction: 'ignore' }))
 
 declare const input: Buffer
 sjson.parse(input)