This PR improves the performance in case malicious payload is provided.

Breaking Change:
- Removed the scan-function

before:

```
node valid
JSON.parse x 1,684,201 ops/sec ±0.74% (90 runs sampled)
JSON.parse proto x 1,090,668 ops/sec ±1.10% (87 runs sampled)
secure-json-parse parse x 1,437,734 ops/sec ±0.97% (88 runs sampled)
secure-json-parse parse proto x 1,580,499 ops/sec ±0.96% (87 runs sampled)
secure-json-parse safeParse x 1,420,574 ops/sec ±0.88% (87 runs sampled)
secure-json-parse safeParse proto x 136,433 ops/sec ±0.78% (90 runs sampled)
JSON.parse reviver x 500,856 ops/sec ±0.58% (89 runs sampled)
Fastest is JSON.parse
```

after:
```
JSON.parse x 1,725,767 ops/sec ±0.75% (91 runs sampled)
JSON.parse proto x 1,089,860 ops/sec ±0.71% (90 runs sampled)
secure-json-parse parse x 1,454,987 ops/sec ±0.73% (93 runs sampled)
secure-json-parse parse proto x 1,570,482 ops/sec ±0.66% (89 runs sampled)
secure-json-parse safeParse x 1,491,839 ops/sec ±0.91% (92 runs sampled)
secure-json-parse safeParse proto x 993,249 ops/sec ±0.95% (87 runs sampled)
JSON.parse reviver x 501,228 ops/sec ±0.79% (92 runs sampled)
Fastest is JSON.parse
```