@mcollina 2022-07-14
Given the implementation of `safeParse()` (which is just a try/catch()), I don't understand why
we had such a big discrepancy in the benchmark to begin with: 1,580,499 vs 136,433. From the look of if you have altered the benchmarks and the tests `secure-json-parse parse` and `secure-json-parse safeParse` are called with different settings. I think it would be better to not check `safeParse` in benchmarks but just focus on the settings of `secure-json-parse parse`.

Note that fastify does not use `safeParse` at all (ref: https://github.com/fastify/fastify/blob/a985b698628fdeeced6d416c4e4f964e2cdb46dd/lib/contentTypeParser.js#L260). Will it be impacted by this change?