@Uzlopak 2022-07-14
@mcollina 

Well instead of throwing it returns null, so no stack trace overhead. 

But if you want i can just create a PR where no active code change exists, but modification on types etc..