@mcollina 2022-07-14
the whole point of `safeParse` vs `parse` is to throw.

Note that the gap between

secure-json-parse safeParse proto x 136,433 ops/sec ±0.78% (90 runs sampled)

and

secure-json-parse safeParse proto x 993,249 ops/sec ±0.95% (87 runs sampled)

cannot be attributed to a try-catch.