@Uzlopak 2022-07-14
Just to be on the same page:
safeParse is about not to throw but to return null