@mcollina 2022-07-14
I'm sorry but I don't understand this sentence:

> Well instead of throwing it returns null, so no stack trace overhead.

You are right that `safeParse()` was always returning `null`. So where is the breaking change?