@Uzlopak 2022-07-14
safeParse runs in the original implementation in a try catch block. 

Now when in safeParse the object is validated to not contain malicious keys, it does not throw a SyntaxError just to be caught in the try catch, but returns null. So no strack trace to be genearated, which gets neglected anyway.