@Uzlopak 2022-07-14
Because I assume it was only exported for unit tests.Also i wanted to avoid unecessary branches