@mcollina 2022-07-14
Consider that Fastify uses `.parse()`: https://github.com/fastify/fastify/blob/main/lib/contentTypeParser.js#L267-L272.